requirements.txt
openai
pandas
pyarrow
pydantic
python-dotenv
duckdb
pyyaml
